requirements.txt
amqp==5.3.1
asgiref==3.11.1
asttokens==3.0.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.9.0
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.6
daphne==4.2.1
decorator==5.2.1
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
executing==2.2.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
matplotlib-inline==0.2.1
msgpack==1.1.2
packaging==26.0
parso==0.8.6
pillow==12.1.1
prompt_toolkit==3.0.52
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
zope.interface==8.2
